+++ src/resolve-with-prefix.test.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { platform } from 'os';
 
-import ResolveWithPrefix from './resolve-with-prefix';
+import { resolveWithPrefix } from './resolve-with-prefix';
 
 const cwd = process.cwd();
 
@@ -13,9 +13,7 @@ test('resolves full preset without prefix set', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix();
-
-    const resolved = resolve('one-preset-test');
+    const resolved = resolveWithPrefix('one-preset-test');
 
     const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
     expect(resolved).toEqual(expected);
@@ -25,9 +23,9 @@ test('resolves full preset with prefix set', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({ prefix: 'one-preset' });
-
-    const resolved = resolve('one-preset-test');
+    const resolved = resolveWithPrefix('one-preset-test', {
+        prefix: 'one-preset',
+    });
 
     const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
     expect(resolved).toEqual(expected);
@@ -37,9 +35,9 @@ test('resolves leading preset with prefix', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({ prefix: 'one-preset' });
-
-    const resolved = resolve('test');
+    const resolved = resolveWithPrefix('test', {
+        prefix: 'one-preset',
+    });
 
     const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
     expect(resolved).toEqual(expected);
@@ -49,9 +47,7 @@ test('resolves other without preset', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix();
-
-    const resolved = resolve('other');
+    const resolved = resolveWithPrefix('other');
 
     const expected = path.resolve(dir, 'node_modules/other/index.js');
     expect(resolved).toEqual(expected);
@@ -61,9 +57,7 @@ test('resolves leading preset with prefix over matching module', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({ prefix: 'one-preset' });
-
-    const resolved = resolve('other');
+    const resolved = resolveWithPrefix('other', { prefix: 'one-preset' });
 
     const expected = path.resolve(
         dir,
@@ -76,12 +70,10 @@ test('strict:true, module: works', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
+    const resolved = resolveWithPrefix('module:explicit-name-test', {
         prefix: 'one-preset',
         strict: true,
     });
-
-    const resolved = resolve('module:explicit-name-test');
 
     const expected = path.resolve(
         dir,
@@ -94,22 +86,16 @@ test('default strict:true, require module:', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
-        prefix: 'one-preset',
-    });
-
     try {
         expect.hasAssertions();
 
-        resolve('explicit-name-test');
+        resolveWithPrefix('explicit-name-test', { prefix: 'one-preset' });
     } catch (error) {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
-        expect(error.message).toMatch(
-            /Cannot find module 'one-preset-explicit-name-test'/,
-        );
-        expect(error.message).toMatch(
-            /If you want to resolve "explicit-name-test", use "module:explicit-name-test"/,
-        );
+        expect(error.message).toMatchInlineSnapshot(`
+            "Cannot find module 'one-preset-explicit-name-test' from '<PROJECT_ROOT>'
+            - If you want to resolve \\"explicit-name-test\\", use \\"module:explicit-name-test\\""
+        `);
     }
 });
 
@@ -117,23 +103,19 @@ test('strict:true, require module:', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
-        prefix: 'one-preset',
-        strict: true,
-    });
-
     try {
         expect.hasAssertions();
 
-        resolve('explicit-name-test');
+        resolveWithPrefix('explicit-name-test', {
+            prefix: 'one-preset',
+            strict: true,
+        });
     } catch (error) {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
-        expect(error.message).toMatch(
-            /Cannot find module 'one-preset-explicit-name-test'/,
-        );
-        expect(error.message).toMatch(
-            /If you want to resolve "explicit-name-test", use "module:explicit-name-test"/,
-        );
+        expect(error.message).toMatchInlineSnapshot(`
+            "Cannot find module 'one-preset-explicit-name-test' from '<PROJECT_ROOT>'
+            - If you want to resolve \\"explicit-name-test\\", use \\"module:explicit-name-test\\""
+        `);
     }
 });
 
@@ -141,21 +123,19 @@ test('strict:true, allow other errors', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
-        prefix: 'one-preset',
-        org: '@example',
-        strict: true,
-    });
-
     try {
         expect.hasAssertions();
-        resolve('one-preset-scoped');
+        resolveWithPrefix('one-preset-scoped', {
+            prefix: 'one-preset',
+            org: '@example',
+            strict: true,
+        });
     } catch (error) {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
-        expect(error.message).toMatch(/Cannot find module 'one-preset-scoped'/);
-        expect(error.message).toMatch(
-            /Did you mean "@example\/one-preset-scoped"/,
-        );
+        expect(error.message).toMatchInlineSnapshot(`
+            "Cannot find module 'one-preset-scoped' from '<PROJECT_ROOT>'
+            - Did you mean \\"@example/one-preset-scoped\\"?"
+        `);
     }
 });
 
@@ -163,11 +143,9 @@ test('strict:true, do not require module: when prefix not set', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
+    const resolved = resolveWithPrefix('explicit-name-test', {
         strict: true,
     });
-
-    const resolved = resolve('explicit-name-test');
 
     const expected = path.resolve(
         dir,
@@ -180,20 +158,18 @@ test('if module not found but it looks like @org missing, add error', () => {
     const dir = path.resolve(__dirname, '__sandbox__/app1/');
     process.chdir(dir);
 
-    const resolve = new ResolveWithPrefix({
-        prefix: 'one-preset',
-        org: '@example',
-    });
-
     try {
         expect.hasAssertions();
-        resolve('one-preset-scoped');
+        resolveWithPrefix('one-preset-scoped', {
+            prefix: 'one-preset',
+            org: '@example',
+        });
     } catch (error) {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
-        expect(error.message).toMatch(/Cannot find module 'one-preset-scoped'/);
-        expect(error.message).toMatch(
-            /Did you mean "@example\/one-preset-scoped"/,
-        );
+        expect(error.message).toMatchInlineSnapshot(`
+            "Cannot find module 'one-preset-scoped' from '<PROJECT_ROOT>'
+            - Did you mean \\"@example/one-preset-scoped\\"?"
+        `);
     }
 });
 
@@ -211,9 +187,10 @@ test('works with NODE_PATH', () => {
         process.env.NODE_PATH = dir;
     }
 
-    const resolve = new ResolveWithPrefix({ prefix: 'one-preset' });
+    const resolved = resolveWithPrefix('one-preset-test', {
+        prefix: 'one-preset',
+    });
 
-    const resolved = resolve('one-preset-test');
     const expected = path.resolve(dir, 'one-preset-test/index.js');
 
     expect(resolved).toEqual(expected);